Fixed dedicated server issue
